added edit functionality

+++ webapp/PMS/src/com/example/pms/CreateCommissionServlet.java
@@ -27,20 +27,20 @@ public class CreateCommissionServlet extends HttpServlet {
 	public void doPost(HttpServletRequest req, HttpServletResponse resp)
 			throws IOException {
 		Double sales = 0.;
-		Integer commRate = 0;
+		Double commRate = 0.;
 		Double prevYearSales = 0.;
 		Double maxSales = 0.;
 		Integer sickDays = 0;
-		Integer yearsWorked = 0;
+		Double yearsWorked = 0.;
 		Double bonus = 0.;
 		
 		try {
 			sales = Double.parseDouble(req.getParameter("sales"));
-			commRate = Integer.parseInt(req.getParameter("commissionRate"));
+			commRate = Double.parseDouble(req.getParameter("commissionRate"));
 			prevYearSales = Double.parseDouble(req.getParameter("previousYearSales"));
 			maxSales = Double.parseDouble(req.getParameter("maxSales"));
 			sickDays = Integer.parseInt(req.getParameter("sickDays"));
-			yearsWorked = Integer.parseInt(req.getParameter("yearsWorked"));
+			yearsWorked = Double.parseDouble(req.getParameter("yearsWorked"));
 			bonus = Double.parseDouble(req.getParameter("bonus"));
 		} catch(NumberFormatException e) {
 			resp.sendRedirect("/createcommissionerror.jsp");
